remove useless ensure
Since whether pushState is supported or not, once route transition in hashchange event and its complete callback is finished, ensureURL will be executed and thus URL will eventually be updated/ensured. Haven't found any edge cases that will prevent from removing this line of code.
If my comprehension is wrong, please add a comment to this line, it confused.
Thanks.